+++ Cargo.toml
@@ -1,13 +1,6 @@
-[package]
-name = "color_generator_rs"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-fast-srgb8 = "1.0.0"
-indicatif = "0.17.7"
-itertools = "0.12.1"
-once_cell = "1.19.0"
-rand = "0.8.5"
+[workspace]
+members = [
+    "lib",
+    "palette-optimizer",
+]
+resolver = "2"
