+++ lib/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod color;
+pub mod metric;
+pub mod score;
+pub mod update;
+
