+++ palette-optimizer/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "palette-optimizer"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+indicatif = "0.17.8"
+itertools = "0.12.1"
+lib = { path = "../lib" }
+rand = "0.8.5"
